requirements.txt
aiogram==3.2.0
aiohttp==3.9.5
python-dotenv==1.0.1
SQLAlchemy==2.0
